requirements.txt
asgiref==3.2.10
Django==3.1
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1